refactor: removed default export from the database feature
