--- apps/backend/src/features/database/index.ts
@@ -1,1 +1,0 @@
-export * from "./kysely";
